Fix syscall_handler argument types in x86_64 cpu.rs
The function is called from assembly using System V AMD64 register
conventions (rdi, rsi, rdx, rcx). On x86_64, usize and u64 are both
64-bit so the change is ABI-safe.

+++ arch/x86_64/src/cpu.rs
@@ -165,6 +165,6 @@ unsafe extern "C" {
 }
 
 #[unsafe(no_mangle)]
-unsafe extern "C" fn syscall_handler(num: u64, arg1: u64, arg2: u64, arg3: u64) -> usize {
+unsafe extern "C" fn syscall_handler(num: usize, arg1: usize, arg2: usize, arg3: usize) -> usize {
     kernel::syscall::handle_syscall(num, arg1, arg2, arg3)
 }
